[Fix] 빌드에러 해결 (#26) - test path에 lombok dependency 제거 - test 코드에 lombok 제거 - RequestDto 분리
